Pay with Amazon Example
A simple example showing Pay with Amazon being used to place an order.
It's rather ugly, but it illustrates the basic principles.  Pay will
Amazon will be incorporated into the cart implementation shortly.

We us postscribe to load amazon dynamically.  The jury is still out on
it.  It seems to work well, but we also saw a few strange errors with
jQuery during the development.  We're not sure we'd use postscribe in
production just yet.  It's not needed if you're willing to hardcode your
Amazon Seller Id in the script src.